% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{generate_name}
\alias{generate_name}
\alias{generate_name_first}
\alias{generate_name_last}
\title{Generate Names}
\usage{
generate_name(n = 10L, sex = c(NA_character_, "f", "m"), year = NULL)

generate_name_first(n = 10L, sex = c(NA_character_, "f", "m"), year = NULL)

generate_name_last(n = 10L)
}
\arguments{
\item{n}{number of samples}

\item{sex}{which sex}

\item{year}{year of birth}
}
\description{
Generate Names
}
