% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomnames-package.R
\docType{data}
\name{names_first}
\alias{names_first}
\title{First names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2020863 rows and 5 columns.
}
\usage{
names_first
}
\description{
First names
}
\keyword{datasets}
