% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomnames-package.R
\docType{data}
\name{names_last}
\alias{names_last}
\title{Last Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 13 columns.
}
\usage{
names_last
}
\description{
Last Names
}
\keyword{datasets}
