% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-names.R
\name{random_name}
\alias{random_name}
\title{Random name generator using adjetive-noun pairs}
\usage{
random_name(n = 1, seed = Sys.time())
}
\arguments{
\item{n}{The number of names desired}

\item{seed}{Optional input to set seed for reproducibility.}
}
\value{
A character vector.
}
\description{
Random name generator using adjetive-noun pairs
}
