% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomnames-package.R
\docType{package}
\name{randomnames-package}
\alias{randomnames}
\alias{randomnames-package}
\title{randomnames: Generate Random Names from US Census}
\description{
Generate random names from US Census. The first names are from 1880-2020. The last names are from 2010.
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
