\name{randomNames-package}
\alias{randomNames-package}
\docType{package}
\title{Function for Generating Random Names and a Dataset
}
\description{
Functions for generating random gender and ethnicity correct first and/or last names where names are proportionally sampled based upon their frequency in
a large scale database.
}

\details{
\tabular{ll}{
Package: \tab randomNames\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2015-7-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
