\name{randomNames-package}
\alias{randomNames-package}
\docType{package}
\title{Functions to create random gender and ethnicity correct first and/or last names where names are proportionally sampled based upon their frequency in
a large scale database.
}
\description{
randomNames contains one function, \code{randomNames} that quickly calculates any number of gender and ethnic specific random first and/or last names.
}

\details{
\tabular{ll}{
Package: \tab randomNames\cr
Type: \tab Package\cr
Version: \tab 0.0-4\cr
Date: \tab 2011-11-19\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
